Add Checkstyle
Checkstyle is now supported on Gradle
