--- src/main/java/Duke/Enums/Commands.java
@@ -1,12 +1,0 @@
-package Duke.Enums;
-
-/**
- * The Commands enum contains a list of all
- * valid commands that Duke accepts
- *
- * @author  Melvin Chan Zijun
- *
- */
-public enum Commands {
-    TODO, EVENT, DEADLINE, LIST, MARK, UNMARK, DELETE, EXIT, HELP, FIND, CLEAR;
-}
