+++ src/main/java/duke/enums/Commands.java
@@ -1,0 +1,12 @@
+package duke.enums;
+
+/**
+ * The Commands enum contains a list of all
+ * valid commands that Duke accepts
+ *
+ * @author  Melvin Chan Zijun
+ *
+ */
+public enum Commands {
+    TODO, EVENT, DEADLINE, LIST, MARK, UNMARK, DELETE, EXIT, HELP, FIND, CLEAR;
+}
